implemented very simple database (now working only for user registration and users list), moved backend stuff in backend folder, updated readme
