+++ backend/package.json
@@ -13,6 +13,7 @@
     "@fastify/static": "^8.1.1",
     "@fastify/swagger": "^9.4.2",
     "fastify": "^5.2.2",
+    "sqlite3": "^5.1.7",
     "uuid": "^11.1.0"
   },
   "devDependencies": {
